requirements.txt
keyboard==0.13.5
Pillow==9.5.0
Pillow==10.2.0
pystray==0.19.5
